reduce Polygons memory usage by 50MB

+++ GlobalQuakeCore/src/main/java/globalquake/core/regions/Region.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import org.geojson.Polygon;
 
-public record Region(String name, List<Path2D.Double> paths, List<Rectangle2D> bounds, List<Polygon> raws) {
+public record Region(String name, List<Path2D.Float> paths, List<Rectangle2D> bounds, List<Polygon> raws) {
 
 
 }
